+++ CHANGES.md
@@ -13,6 +13,8 @@
 * Added JTableRowIndexToViewTransformer
 * Added JTableColumnIndexToModelTransformer
 * Added JTableColumnIndexToViewTransformer
+* Moved CollectionElementTransformer to 'collection' sub-package
+* Added GetCollectionSizeTransformer
 * Fixed issue where the value of the JTextComponentTextProperty was not immediately updated upon replacement of the document
 
 # Version 3.2.6